+++ runtime/src/test/java/com/xatkit/plugins/react/platform/io/ReactIntentProviderTest.java
@@ -1,17 +1,19 @@
 package com.xatkit.plugins.react.platform.io;
 
+import com.xatkit.AbstractXatkitTest;
 import com.xatkit.XatkitTest;
 import com.xatkit.core.XatkitCore;
 import com.xatkit.plugins.react.platform.ReactPlatform;
 import com.xatkit.stubs.StubXatkitCore;
 import org.apache.commons.configuration2.BaseConfiguration;
 import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
 import static java.util.Objects.nonNull;
 
-public class ReactIntentProviderTest extends XatkitTest {
+public class ReactIntentProviderTest extends AbstractXatkitTest {
 
     private static XatkitCore xatkitCore;
 
@@ -23,6 +25,14 @@ public class ReactIntentProviderTest extends XatkitTest {
     public static void setUpBeforeClass() {
         xatkitCore = new StubXatkitCore();
         reactPlatform = new ReactPlatform(xatkitCore, new BaseConfiguration());
+    }
+
+    @AfterClass
+    public static void tearDownAfterClass() {
+        reactPlatform.shutdown();
+        if(!xatkitCore.isShutdown()) {
+            xatkitCore.shutdown();
+        }
     }
 
     @After