+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 
 ## Unreleased
 
+### Added
+
+- Configuration property `xatkit.react.client.url` to specify the URL of client domains embedding the react chat component.  This property is translated into a `Access-Control-Allow-Origin` header that tells the browser to authorize requests from the specified domain. Supported values for the property are single URL and wildcard (`"*"`). The standard does not authorize multiple URLs in the `Access-Control-Allow-Origin` header (see [here]( https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin )).
+
 ## [3.0.0]- 2019-10-10
 
 ### Added